Refactor Normalize() in the command pkg
