+++ internal/command/command_args_normalizer.go
@@ -9,6 +9,7 @@ import (
 	"strings"
 
 	"github.com/pkg/errors"
+	"go.uber.org/multierr"
 	"go.uber.org/zap"
 	"google.golang.org/protobuf/proto"
 
@@ -40,7 +41,11 @@ type argsNormalizer struct {
 	scriptFile       *os.File
 }
 
-func (n *argsNormalizer) Normalize(cfg *Config) (*Config, error) {
+func newArgsNormalizer(session *Session, logger *zap.Logger) configNormalizer {
+	return (&argsNormalizer{session: session, logger: logger}).Normalize
+}
+
+func (n *argsNormalizer) Normalize(cfg *Config) (*Config, func() error, error) {
 	args := append([]string{}, cfg.Arguments...)
 
 	switch cfg.Mode {
@@ -51,7 +56,7 @@ func (n *argsNormalizer) Normalize(cfg *Config) (*Config, error) {
 
 		if isShellLanguage(filepath.Base(cfg.ProgramName)) {
 			if err := n.inlineShell(cfg, &buf); err != nil {
-				return nil, err
+				return nil, nil, err
 			}
 		} else {
 			// Write the script from the commands or the script.
@@ -65,21 +70,22 @@ func (n *argsNormalizer) Normalize(cfg *Config) (*Config, error) {
 			}
 		}
 
+		// TODO(adamb): "-c" is not supported for all inline programs.
 		if val := buf.String(); val != "" {
 			args = append(args, "-c", val)
 		}
 
 	case *runnerv2alpha1.CommandMode_COMMAND_MODE_FILE.Enum():
 		if err := n.createTempDir(); err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 
 		if err := n.createScriptFile(); err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 
 		if err := n.writeScript([]byte(cfg.GetScript())); err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 
 		// TODO(adamb): it's not always true that the script-based program
@@ -89,7 +95,7 @@ func (n *argsNormalizer) Normalize(cfg *Config) (*Config, error) {
 
 	result := proto.Clone(cfg).(*Config)
 	result.Arguments = args
-	return result, nil
+	return result, n.cleanup, nil
 }
 
 func (n *argsNormalizer) inlineShell(cfg *Config, buf *strings.Builder) error {
@@ -125,7 +131,17 @@ func (n *argsNormalizer) inlineShell(cfg *Config, buf *strings.Builder) error {
 	return nil
 }
 
-func (n *argsNormalizer) Cleanup() error {
+func (n *argsNormalizer) cleanup() (result error) {
+	if err := n.collectEnv(); err != nil {
+		result = multierr.Append(result, err)
+	}
+	if err := n.removeTempDir(); err != nil {
+		result = multierr.Append(result, err)
+	}
+	return
+}
+
+func (n *argsNormalizer) removeTempDir() error {
 	if n.tempDir == "" {
 		return nil
 	}
@@ -139,7 +155,7 @@ func (n *argsNormalizer) Cleanup() error {
 	return nil
 }
 
-func (n *argsNormalizer) CollectEnv() error {
+func (n *argsNormalizer) collectEnv() error {
 	if n.session == nil || !n.isEnvCollectable {
 		return nil
 	}
@@ -236,23 +252,17 @@ func splitNull(data []byte, atEOF bool) (advance int, token []byte, err error) {
 	return 0, nil, nil
 }
 
-type envSource func() []string
+func shellOptionsFromProgram(programPath string) (res string) {
+	base := filepath.Base(programPath)
+	shell := base[:len(base)-len(filepath.Ext(base))]
 
-type envNormalizer struct {
-	sources []envSource
-}
-
-func (n *envNormalizer) Normalize(cfg *Config) (*Config, error) {
-	result := proto.Clone(cfg).(*Config)
-
-	env := os.Environ()
-	env = append(env, cfg.Env...)
-
-	for _, source := range n.sources {
-		env = append(env, source()...)
+	// TODO(mxs): powershell and DOS are missing
+	switch shell {
+	case "zsh", "ksh", "bash":
+		res += "set -e -o pipefail"
+	case "sh":
+		res += "set -e"
 	}
 
-	result.Env = env
-
-	return result, nil
+	return
 }
